STM32F4 - flash_api - cache cleaning fix
Seems the F4 family is also missing the cache cleanup fixes in the flash_api.c.
Same fix as in original Mbed OS PR by Vincent Veron <vincent.veron@st.com>

https://github.com/ARMmbed/mbed-os/pull/10248

However, as the cache invalidation instructions are not available in CM4,
we're trying to use the FLASH_FlushCaches available in stm32f4xx_hal_flash.
(STM has implemented their own flash cache).

In case of programming the flash we should also flush the caches, as it is
theoretically possible the programming failed but a readback test would not
find it, if it gets that readback from the cache, instead of the flash.
Thus a potential write failure could go unnoticed.