Update TDBStore to Mbed OS 6.3 level (2f87d59c7f)
Copy over the TDBStore.cpp|.h as they contain a few bug fixes which we seem
to be needing (in some corner cases). The files are copied over from Mbed OS master
at the tip of master as of today, hash pointing at 2f87d59c7f.

This brings in a number of changes, but primarily we are interested in this:
https://github.com/ARMmbed/mbed-os/pull/11987

Device 943 ended up in a state, where it would not boot up and in the debug
image I was able to spot it died in the TDBStore.cpp,

00> ++ MbedOS Error Info ++
00> Error Status: 0x80FF011B Code: 283 Module: 255
00> Error Message: TDBSTORE: Unable to build RAM table at init
00> Location: 0x71347
00> Error Value: 0x0
00> Current Thread: main Id: 0x20012C68 Entry: 0x7482F StackSize: 0x1400 StackMem: 0x200168E8 SP: 0x20017AEC
00> For more info, visit: https://mbed.com/s/error?error=0x80FF011B&tgt=EPEN3

This is due to corrupted QSPIF. This should prevent that from happening or at least
recover from it in a more sensible way, as it will not "guess if it's been erased or not",
but more deterministily just erase properly.

Device was recoverable even with existing SW by forcing an erase in the error branch,
patch below:

--- start of patch:

```
diff --git i/features/storage/kvstore/tdbstore/TDBStore.cpp w/features/storage/kvstore/tdbstore/TDBStore.cpp
index efeedf42c2..701ce7d044 100644
--- i/features/storage/kvstore/tdbstore/TDBStore.cpp
+++ w/features/storage/kvstore/tdbstore/TDBStore.cpp
@@ -1027,6 +1027,7 @@ int TDBStore::init()
     }

 #endif
+    printf("_bd.type == %s\n", _bd->get_type());

     _max_keys = initial_max_keys;

@@ -1097,6 +1098,7 @@ int TDBStore::init()

     // In case we have two empty areas, arbitrarily use area 0 as the active one.
     if ((area_state[0] == TDBSTORE_AREA_STATE_EMPTY) && (area_state[1] == TDBSTORE_AREA_STATE_EMPTY)) {
+        printf("active area = 0\n");
         _active_area = 0;
         _active_area_version = 1;
         ret = write_master_record(_active_area, _active_area_version, _free_space_offset); @@ -1115,6 +1117,7 @@ int TDBStore::init()
         } else {
             _active_area = 1;
         }
+        printf("active area = %d\n", _active_area);
         _active_area_version = versions[_active_area];
         ret = reset_area(1 - _active_area);
         if (ret) {
@@ -1128,6 +1131,11 @@ int TDBStore::init()
     ret = build_ram_table();

     if ((ret != MBED_SUCCESS) && (ret != MBED_ERROR_INVALID_DATA_DETECTED)) {
+        printf("TDBSTORE: Unable to build RAM table at init - try erase area!");
+        ret = reset_area(0);
+        printf("reset_area(0))=%d\n", ret);
+        ret = reset_area(1);
+        printf("reset_area(1))=%d\n", ret);
         MBED_ERROR(ret, "TDBSTORE: Unable to build RAM table at init");
     }

@@ -1159,6 +1167,7 @@ int TDBStore::init()
     }

 end:
+    printf("TDBStore init() done for %s\n", _bd->get_type());
     _is_initialized = true;
     _mutex.unlock();
     return ret;
```
--- end of patch